Moved definition of date and time format into webinterface
Move DATE_FORMAT and TIME_FORMAT to webinterface.
Also modified Zoph in a few places so that this setting is used
for more places where time or date is displayed. Also got rid of a few
date/time related functions in util.inc.php by replacing the with
calls to the Time class.

#28