Moved settings from settings.inc.php to conf.inc.php
The settings object was a built as a solution to hold settings that
could be changed at runtime (contrary to the defines that were used
for config, which cannot change once defined).

Initially, my plan was to base the webbased configuration on this
object, but I decided I didn't like the way that would turn out and
built a new infrastructure to hold settings. This infrastructure is
now there in the form of the conf and confItem* object.

Since it's odd to have two seperate infrastructures for holding
configuration, it's time to start phasing "settings" out.

- Introduced confItemNumber and migrated a few settings
  to the new type.
- Added the possibility to mark a configuration item "internal",
  this means it will not be exposed to the webinterface.
- Moved the settings in settings.inc.php to conf.inc.php
- In the near future the settings object will change to the ini object.

#28