Removed several global variables
Removed a couple of global variables:
- Moved $DEFAULT_ORDER and $DEFAULT_DIRECTION to web configuration
- The others were part of user preferences, have removed the need for
  using global variables there.

Created a static variable in the user class to register the currently
logged on user, so we can get rid of the global $user variable as well
as functions that pass $user around ad nauseum. So far, this has only
been done in the breadcrumbs include, but it should be relatively easy
to do anywhere else.

Changed the translate method of language so that it can also handle
arrays, so we can easily translate an entire array.

#28