Removed all migrated config from config.inc.php
A small list of configuration-items remain:

VERSION - Zoph internal so I don't need to change the version number all
over the place.

INI_FILE - Zoph needs to find the ini file before it can go to the
database, so this setting can not be moved into database.

THUMB_SIZE, MID_SIZE, THUMB_PREFIX, MID_PREFIX - In a future version it
will be possible to define these from the webinterface, but this will
require a major overhaul in many parts of Zoph.

LOG_ALWAYS, LOG_SEVERITY, LOG_SUBJECT - These are debug-related settings
that could cause troubles when changing them in the database and also,
it would make debugging of the part where the database is not yet
initialized impossible.

Closes #28