Modified internal database references to static
Changed the internal variables that Zoph uses to figure out which
database table belongs to which object (as well as the primary keys and
fields that cannot be null) from object variables into static variables.

Also modified the getRecords(), getRecordsFromQuery() and
getTopNfromSQL() and  method to no longer need the class name as a
parameter.

All of the above became possible because Zoph no longer supports PHP
5.2, as "Late Static Binding", which was needed for this change, was
introduced in PHP 5.3.