CLI tries to lookup previous argument's value when looking up photographer
A bug in the CLI caused Zoph to try to lookup the value for the argument
before --photographer instead of the value after it.

Also fixed an issue where the CLI was trying to find the CLI verbosity
setting in the "old" location (settings object) instead of the new
location (conf object).

Fixes #38