Refactor getXML
Made several changes to the way XML data is retrieved, instead of one
big function with a lot of `if ($object == "xxx")` statements, it's now
a nice OO function.

- Refactored the global get_people_xxx functions in the person and
photographer object into static functions. This makes some changes in
the way data is retrieved in some places, photographer::getAll(), for
example now returns all persons when called by an admin user. This
should not have any effect on the way Zoph works, but it greatly
simplifies the internals.

- Changed the xml_rootname()/xml_nodename() methods in several classes
into a class constants XMLROOT and XMLNODE.