Some small usability improvements
On the album screen, moved the "edit" and "remove coverphoto"
actionlinks into the header, this is consistent with the rest of Zoph
and makes it possible to enter the edit page even if the normal display
is overrided by a pageset with "show original" set to "never".

On the person screen, the name of the person in shown in the title
instead of "person".