Made changes to support MySQL STRICT mode support
MySQL STRICT mode, which is the default for new installation as of MySQL
v5.6.8, does not allow inserting a string that is longer than the field
size into a field or inserting the string "null" into a integer field.

This change modifies the parts of Zoph that were sinning against this
rule.

Fixes #49