Added changing the autocover selection to circles
Added the same selection criteria for autocover selection as albums,
categories, places and people: oldest, newest, first, last, random,
highest. BUT circles always FIRST gets the photo with the most people on
it. So, if you have set autocover to "highest" (= highest rating) and
you have a circle with 5 people in it, and have 3 photos:
1. rating = 5, 4 of the 5 people are on it
2. rating = 6, 4 of the 5 people are on it, and
3. rating = 10, 3 of the 5 people are on it
It will select photo 2.

Issue #44