Making database changes for Issue #46
I am adding the possibility to hide members of a certain circle. This is
mainly meant to keep overview, even with a large database, with 100s of
people of which many may have been photographed only once or a few
times.

This commit makes prepares the database for this change.

Issue #46