Hide people with no photos in Circle view
In 'classic' people view, people that were not on a photo and had not
taken a photo - in either case that means: a photo the user is allowed
to see, these people were hidden for that user as well.

Until now, those people were visible in the 'circles' view. This has now
changed and the behaviour is consistent.

Issue #46