Fixed unittests
It turned out that the script creating the testdata contained an issue
that caused an admin user not to be defined as such. Since this user was
used in a lot of unittests, these were all failing after fixing this
issue. This was fixed by changing the unittests to use a different user.