Modified query for photo access rights to a view for performance reasons
Because of the way Zoph's permissions system works, determining which
photos, categories, people, albums and places a specific user is allowed
to see is a non-trivial excercise. It involves several tables, and thus
a complex query consisting of serveral JOINs. The way the query was
built was not working very well with MySQL's query optimizer, and caused
some of these queries to take up to 2 seconds to run (with a large
database). If a page executes several of these queries, browsing through
Zoph can because very slow. Especially the people page suffered from
this problem.

This change adds a view to Zoph, which simplifies the queries a lot,
since all the complexity is hidden in the view, but also enables the
MySQL query optimizer to do a much better job to retrieve the same data.
The above-mentionned queries that took 2 seconds are now down to about
150ms...