Performance improvement on place page
The place page could take a long time when loading, especially for non-
admin users. In this changed I implemented the same solution as I
previously implemented in categories.

I took the code from categories and made it reusable and modified places
to use it too.

The "remove_empty()" function is now no longer used, so it was removed.
(It turned out not to be a very good idea anyway)

I also removed the ancestorCache since it didn't work very well and
didn't cause any noticable improvement in performance.