Added requirements to configuration
It is now possible to disable a configuration item based on whether or
not another configuration item is enabled.

This was needed for the logon background album selection, since it
requires the sharing feature to be enabled.

At this moment the function is only implemented for SELECT boxes as that
is the one currently needed, it should be easy to expand it to other
type of configuration items once the need is there.

Fixes #72