Changed Zoph to PHP's password functions instead of using MySQL's password()
The new function is more secure due to random salting (i.e. two users
with the same password will not have the same hash).

The hashes in the database will automatically be updated when the user
logs in for the first time. The migration mechanism will eventually be
removed.

Fixes #75