Modified access checking to enable users to add photos to albums they have created
Since the access checking works by checking for photos a user can see,
an empty album is by definition invisible for a non-Admin (and non-'can
see all photos'-user). This means that a user can create an album, but
consequently, it's invisble for that user. This change adds a creator to
each album, and allows the creator to see the album even without photos
in it.
This change also adds the basic infrastructure to do the same for
categories, places, people and circles. But does not yet expose this to
the outside.

Issue #22