Modify recursive creation of directories
Changed the order in which recursive creation of directories is
executed. The 'old' way was to start at the root and work the way up to
the dir to be created. This caused problems when open_basedir
restriction is activated. The 'new' behaviour is to start at the dir to
be created, then work up to the first dir that needs to be created and
then work back down. As long as the "basedir" exists, this will keep
Zoph from trying to pry on the filesystem outside of the basedir.

Also, did some refactoring, moving a few related function out of
util.inc.php into the file class.

Issue #79