Issue#83: Very crude first implementation of photo collection object
To cleanup the search page, a new search backend first needs to be
created. The search backend now consists of a huge global function
get_photos(). This is a first implementation of a photo collection
object to replace that function. This first implementation is only a
wrapper around the old function.

Issue #83