Issue#100 Deprecate ssl.force, url.http and url.https
Introduced the possibility to deprecate configuration items, display a
warning on the configuration page and if an item is used (that is: set
to something different then the default) a warning is displayed on the
main Zoph page.

In the next version these items will be removed. Please file a bug if
you see these warnings!

Fixes #100